+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog], and this project adheres to [Semantic Versioning].
 
+## Unreleased
+
+### Changed
+
+- Switch deprecated ghostnet to shadownet as default network.
+
 ## [3.17.0](https://github.com/baking-bad/pytezos/compare/3.16.1...3.17.0) - 2026-01-19
 
 ### Added